Fixed App.test and moved tests to a separate folder inside src
