+++ src/tests/RouteList.test.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import { render } from '@testing-library/react';
-import RouteList from './RouteList';
+import RouteList from '../components/routeList/RouteList';
 
 test('renders learn react link', () => {
   const { getByText } = render(<RouteList />);
